+++ src/main/java/com/apocollis/aqtweaks/ClientProxy.java
@@ -4,6 +4,7 @@ import com.apocollis.aqtweaks.client.ClientModule;
 import com.apocollis.aqtweaks.depths.DepthsFogHandler;
 import com.apocollis.aqtweaks.portal.EntityArcaneRift;
 import com.apocollis.aqtweaks.portal.client.RenderArcaneRift;
+import com.apocollis.aqtweaks.stamina.DssSkillsGuiClient;
 import com.apocollis.aqtweaks.stamina.StaminaModuleClient;
 
 import net.minecraftforge.common.MinecraftForge;
@@ -26,6 +27,7 @@ public class ClientProxy extends CommonProxy {
     public void init(FMLInitializationEvent event) {
         super.init(event);
         MinecraftForge.EVENT_BUS.register(new StaminaModuleClient());
+        DssSkillsGuiClient.register();
         MinecraftForge.EVENT_BUS.register(new DepthsFogHandler());
         MinecraftForge.EVENT_BUS.register(new ClientModule());
     }